+++ sdk/ml/azure-ai-ml/azure/ai/ml/_schema/workspace/connections/credentials.py
@@ -4,6 +4,10 @@
 
 # pylint: disable=unused-argument
 
+##### DEV NOTE: For some reason, these schemas correlate to the classes defined in ~azure.ai.ml.entities._credentials.
+# There used to be a credentials.py file in ~azure.ai.ml.entities.workspace.connections,
+# but it was, as far as I could tell, never used. So I removed it and added this comment.
+
 from typing import Dict
 
 from marshmallow import fields, post_load