+++ sdk/ml/azure-ai-ml/azure/ai/ml/constants/_common.py
@@ -186,6 +186,8 @@ SPARK_ENVIRONMENT_WARNING_MESSAGE = (
 CONNECTION_API_VERSION_KEY = "ApiVersion"
 CONNECTION_API_TYPE_KEY = "ApiType"
 CONNECTION_KIND_KEY = "Kind"
+CONNECTION_CONTAINER_NAME_KEY = "ContainerName"
+CONNECTION_ACCOUNT_NAME_KEY = "AccountName"
 
 
 class DefaultOpenEncoding:
@@ -791,9 +793,10 @@ class InputOutputModes:
 
 
 class WorkspaceConnectionTypes:
-    """Names for workspace connection types that are different from that underlying api enum values."""
+    """Names for workspace connection types that are different from that underlying api enum values
+    from the ConnectionCategory class."""
 
-    CUSTOM = "custom"
+    CUSTOM = "custom"  # Corresponds to "custom_keys".
 
 
 class LegacyAssetTypes: