+++ sdk/ml/azure-ai-ml/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Features Added
 - Remove `experimental` tag for  `ml_client.jobs.validate`.
+- Workspace Connection has new read-only subclass: AzureBlobStoreWorkspaceConnectionSchema.
+- Workspace Connection supports 2 new types under main class: gen 2 and azure_one_lake.
+- Workspace Connection LIST operation can return data connections via new optional flag: include_data_connections.
 
 ### Bugs Fixed
 - Fix pipeline job `outputs` not load correctly when `component: <local-file>` exists in pipeline job yaml.